Handled a case where table matrix is empty
For some reason React builds table at least twice now when matrix
is missing. This might be because dispatcher messaging messes
something up or the asynchronicity of the data request
allows React to go ahead and render when app is not in correct state.

Regards of this, loading table data from server works.
Loading table metadata is still to come.