Managed to get full table loading to work
With a table of 20K cells, about 150 rows and cols, the speed
is already too slow (as was to be expected), and stability flaky
(which I hadn't really given much thought yet).
This is a good result with everything going, like debug,
hot reload, the works. So, it is essentially unoptimized.